requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.5
pymongo>=4.6.0
mongoengine>=0.27.0
Pillow>=10.3.0
boto3==1.34.34
django-storages==1.14.2
python-decouple==3.8
celery==5.3.4
redis==5.0.1
django-redis==5.4.0
drf-spectacular==0.27.0
setuptools>=65.0.0

